[Feat] Auth 서버 Feign 요청 및 token 파싱

+++ src/main/java/org/click/friends/service/FriendServiceImpl.java
@@ -5,51 +5,79 @@ import lombok.RequiredArgsConstructor;
 import org.click.friends.dto.request.ConfirmFriendRequest;
 import org.click.friends.dto.request.FriendRequest;
 import org.click.friends.entity.Friend;
+import org.click.friends.global.api.ApiUser;
+import org.click.friends.global.dto.UserListResponse;
+import org.click.friends.global.dto.UserResponse;
 import org.click.friends.repository.FriendRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
 public class FriendServiceImpl implements FriendService {
 
     private final FriendRepository friendRepository;
+    private final ApiUser apiUser;
 
     // 친구 목록 조회
     @Override
-    public List<Friend> getFriends(UUID myId) {
-        return friendRepository.findByFriendshipIsTrueAndMyId(myId);
+    public List<UserListResponse> getFriends(String myCode) {
+        List<Friend> friendList = friendRepository.findByFriendshipIsTrueAndMyCode(myCode);
+
+        String[] collect = friendList.stream()
+                .map(Friend::getTargetCode)
+                .toArray(String[]::new);
+
+        return apiUser.getUsers(collect);
     }
 
     // 친구 요청
     @Override
-    public void acceptFriendRequest(FriendRequest request) {
+    public void acceptFriendRequest(String code, String myCode) {
+        UserResponse user = apiUser.getUser(code);
+
+        if(user == null)
+            throw new IllegalArgumentException("유저를 찾을 수 없습니다.");
+        FriendRequest request = new FriendRequest(null, false, code);
         // 이미 친구 요청을 보낸 유저입니다.
-        if(friendRepository.existsByFriendshipIsFalseAndMyIdAndTargetId(request.my_id(), request.target_id()))
+        if(friendRepository.existsByFriendshipIsFalseAndMyCodeAndTargetCode(myCode, code)
+        || friendRepository.existsByFriendshipIsFalseAndMyCodeAndTargetCode(code, myCode))
             throw new IllegalArgumentException("이미 친구 요청을 보낸 유저입니다.");
         // 이미 친구 사이인 유저입니다.
-        if(friendRepository.existsByFriendshipIsTrueAndMyIdAndTargetId(request.my_id(), request.target_id()))
+        if(friendRepository.existsByFriendshipIsTrueAndMyCodeAndTargetCode(myCode, code))
             throw new IllegalArgumentException("이미 친구 사이인 유저입니다.");
-        friendRepository.save(request.toEntity());
+        // 자기 자신입니다.
+        if(code.equals(myCode))
+            throw new IllegalArgumentException("자기 자신입니다.");
+
+        friendRepository.save(request.toEntity(myCode));
     }
 
     // 친구 요청 수락
     @Override
     @Transactional
-    public void confirmFriendRequest(Long friendId, ConfirmFriendRequest request) {
-        Friend friend = friendRepository.findById(friendId).orElseThrow(IllegalArgumentException::new);
+    public void confirmFriendRequest(String code, String myCode) {
+        UserResponse user = apiUser.getUser(code);
+        if(user == null)
+            throw new IllegalArgumentException("유저를 찾을 수 없습니다.");
+
+        ConfirmFriendRequest request = new ConfirmFriendRequest(null, true, code);
+        Friend friend = friendRepository.findByFriendshipIsFalseAndMyCodeAndTargetCode(code, myCode);
         friend.setFriendship(true);
 
         friendRepository.save(friend); // 기존 데이터 업데이트
-        friendRepository.save(request.toEntity()); // user_id 순서 바꿔서 새로 저장
+        friendRepository.save(request.toEntity(myCode)); // user_id 순서 바꿔서 새로 저장
     }
 
     // 친구 요청 삭제
     @Override
-    public void rejectFriendRequest(Long friendId) {
-        Friend friend = friendRepository.findById(friendId).orElseThrow(IllegalArgumentException::new);
+    public void rejectFriendRequest(String code, String myCode) {
+        UserResponse user = apiUser.getUser(code);
+        if(user == null)
+            throw new IllegalArgumentException("유저를 찾을 수 없습니다.");
+
+        Friend friend = friendRepository.findByFriendshipIsFalseAndMyCodeAndTargetCode(code, myCode);
 
         friendRepository.delete(friend);
     }
@@ -57,16 +85,24 @@ public class FriendServiceImpl implements FriendService {
     // 친구 삭제
     @Override
     @Transactional
-    public void removeFriend(Long friendId) {
-        Friend friend = friendRepository.findById(friendId).orElseThrow(IllegalArgumentException::new);
+    public void removeFriend(String code, String myCode) {
+        UserResponse user = apiUser.getUser(code);
+        if(user == null)
+            throw new IllegalArgumentException("유저를 찾을 수 없습니다.");
 
-        friendRepository.deleteByMyIdAndTargetId(friend.getTargetId(), friend.getMyId());
-        friendRepository.delete(friend);
+        friendRepository.deleteByMyCodeAndTargetCode(myCode, code);
+        friendRepository.deleteByMyCodeAndTargetCode(code, myCode);
     }
 
     // 친구 요청 목록
     @Override
-    public List<Friend> getFriendRequests(UUID myId) {
-        return friendRepository.findByFriendshipIsFalseAndTargetId(myId);
+    public List<UserListResponse> getFriendRequests(String myCode) {
+        List<Friend> requestFriendList = friendRepository.findByFriendshipIsFalseAndTargetCode(myCode);
+
+        String[] collect = requestFriendList.stream()
+                .map(Friend::getMyCode)
+                .toArray(String[]::new);
+
+        return apiUser.getUsers(collect);
     }
 }
